Use arch_flags instead of use_... flags
This should make it easier to add some more thorough testing of configs.

I also think this is worth doing just because it's not good to have the same information stored in two different ways. I already found some flags that were missing when making the previous change in this sequence of changes.

PiperOrigin-RevId: 769779472